# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LedgerOS accounting system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: mine
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Submit expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}/transition:
    post:
      operationId: transitionExpense
      tags: [expenses]
      summary: Review expense (submit, approve, reject, request-receipt, non-reimbursable)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionInput"
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /monthly-close:
    get:
      operationId: listMonthlyClose
      tags: [close]
      summary: List monthly close periods
      responses:
        "200":
          description: Close periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyClose"
    post:
      operationId: createMonthlyClose
      tags: [close]
      summary: Create a monthly close period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonthlyCloseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyClose"
  /monthly-close/{id}/checklist:
    post:
      operationId: updateCloseChecklist
      tags: [close]
      summary: Toggle a checklist item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistUpdate"
      responses:
        "200":
          description: Close period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyClose"
  /monthly-close/{id}/lock:
    post:
      operationId: lockMonthlyClose
      tags: [close]
      summary: Lock the period (approval required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Close period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyClose"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TransitionInput:
      type: object
      required: [action]
      properties:
        action: { type: string, minLength: 1 }
        note: { type: string }
        matchBankTransactionId: { type: ["integer", "null"] }
...
    Customer:
      type: object
      required: [id, name, balance]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        sourceApp: { type: ["string", "null"] }
        externalId: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        balance: { type: number }
        isTestData: { type: boolean }
    CustomerInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        company: { type: string }
        sourceApp: { type: string }
        externalId: { type: string }
        notes: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        company: { type: string }
        sourceApp: { type: string }
        notes: { type: string }
...
    Expense:
      type: object
      required: [id, description, amount, status, expenseDate]
      properties:
        id: { type: integer }
        description: { type: string }
        vendorName: { type: ["string", "null"] }
        amount: { type: number }
        category: { type: ["string", "null"] }
        clientProject: { type: ["string", "null"] }
        sourceApp: { type: ["string", "null"] }
        status: { type: string }
        expenseDate: { type: string }
        hasReceipt: { type: boolean }
        reimbursable: { type: boolean }
        submittedById: { type: ["integer", "null"] }
        submittedByName: { type: ["string", "null"] }
        isTestData: { type: boolean }
        createdAt: { type: string }
    ExpenseInput:
      type: object
      required: [description, amount, expenseDate]
      properties:
        description: { type: string, minLength: 1 }
        vendorName: { type: string }
        amount: { type: number }
        category: { type: string }
        clientProject: { type: string }
        sourceApp: { type: string }
        expenseDate: { type: string }
        hasReceipt: { type: boolean }
...
    ChecklistItem:
      type: object
      required: [key, label, done]
      properties:
        key: { type: string }
        label: { type: string }
        done: { type: boolean }
    MonthlyClose:
      type: object
      required: [id, periodLabel, status, checklist]
      properties:
        id: { type: integer }
        periodLabel: { type: string }
        periodStart: { type: string }
        periodEnd: { type: string }
        status: { type: string }
        lockedAt: { type: ["string", "null"] }
        lockedById: { type: ["integer", "null"] }
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        isTestData: { type: boolean }
    MonthlyCloseInput:
      type: object
      required: [periodLabel, periodStart, periodEnd]
      properties:
        periodLabel: { type: string, minLength: 1 }
        periodStart: { type: string }
        periodEnd: { type: string }
    ChecklistUpdate:
      type: object
      required: [key, done]
      properties:
        key: { type: string }
        done: { type: boolean }
...